Added very basic user input.

+++ src/FRBC_Coding_Assignment/ConsoleApp.cs
@@ -1,5 +1,6 @@
 ﻿using FRBC_Coding_Assignment.Services.Interfaces;
 using System;
+using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 
 namespace FRBC_Coding_Assignment
@@ -12,6 +13,7 @@ namespace FRBC_Coding_Assignment
         private readonly IPorterStemmingService porterStemmingService;
         private readonly IWordOccurrenceService wordOccurrenceService;
         private readonly IStopWordService stopWordService;
+        private readonly Dictionary<string, string> fileNames = new Dictionary<string, string> { { "1", "Text1" }, { "2", "Text2"} };
 
         public ConsoleApp(
             IFileService fileService,
@@ -30,29 +32,56 @@ namespace FRBC_Coding_Assignment
 
         public void Run()
         {
-            // Read a file
-            var dirtyText = fileService.ReadAllTextFromFile("Text2");
-            var nonUnicodeText = stringCleanerService.RemoveUnicodeCharacters(dirtyText);
+            try
+            {
+                // Get User Input
+                Console.WriteLine("Please pick one:");
+                Console.WriteLine("1. Text1.txt");
+                Console.WriteLine("2. Text2.txt");
 
-            // Remove Stop Words
-            var noPunctuationOrSymbols = stringCleanerService.RemovePunctuationAndSymbols(nonUnicodeText);
-            var semiSanitizedArray = noPunctuationOrSymbols.Split(" ", StringSplitOptions.RemoveEmptyEntries);
-            var sanitizedList = stringCleanerService.RemoveApostropheExceptConjunctions(semiSanitizedArray);
-            var stopWordRemovedList = stopWordService.RemoveStopWords(sanitizedList);
+                string fileChoice = Console.ReadLine();
+                if (fileChoice.Equals("1") || fileChoice.Equals("2"))
+                {
+                    // Read a file
+                    var dirtyText = fileService.ReadAllTextFromFile(fileNames[fileChoice]);
+                    var nonUnicodeText = stringCleanerService.RemoveUnicodeCharacters(dirtyText);
 
-            // Remove all non alpha characters
-            var alphaOnlyText = stringCleanerService.RemoveNonAlphaCharacters(string.Join(" ", stopWordRemovedList));
+                    // Remove Stop Words
+                    var noPunctuationOrSymbols = stringCleanerService.RemovePunctuationAndSymbols(nonUnicodeText);
+                    var semiSanitizedArray = noPunctuationOrSymbols.Split(" ", StringSplitOptions.RemoveEmptyEntries);
+                    var sanitizedList = stringCleanerService.RemoveApostropheExceptConjunctions(semiSanitizedArray);
+                    var stopWordRemovedList = stopWordService.RemoveStopWords(sanitizedList);
 
-            // Stemming Algorithm
-            var alphaOnlyArray = alphaOnlyText.Split(" ", StringSplitOptions.RemoveEmptyEntries);
-            var stemmingArray = porterStemmingService.RunStemmingAlgorithm(alphaOnlyArray);
+                    // Remove all non alpha characters
+                    var alphaOnlyText = stringCleanerService.RemoveNonAlphaCharacters(string.Join(" ", stopWordRemovedList));
 
-            // Computes the frequency of each term
-            var wordOccurrenceCount = wordOccurrenceService.GetWordOccurences(stemmingArray);
-            var sortedOccurrenceCount = wordOccurrenceService.SortWordOccurences(wordOccurrenceCount);
+                    // Stemming Algorithm
+                    var alphaOnlyArray = alphaOnlyText.Split(" ", StringSplitOptions.RemoveEmptyEntries);
+                    var stemmingArray = porterStemmingService.RunStemmingAlgorithm(alphaOnlyArray);
 
-            // Prints out the most commonly occurring terms (not including stop words) in descending order of frequency
-            wordOccurrenceService.PrintTopWordOccurences(sortedOccurrenceCount);
-        }
+                    // Computes the frequency of each term
+                    var wordOccurrenceCount = wordOccurrenceService.GetWordOccurences(stemmingArray);
+                    var sortedOccurrenceCount = wordOccurrenceService.SortWordOccurences(wordOccurrenceCount);
+
+                    // Prints out the most commonly occurring terms (not including stop words) in descending order of frequency
+                    wordOccurrenceService.PrintTopWordOccurences(sortedOccurrenceCount);
+                }
+                else
+                {
+                    Console.WriteLine("Input entered wrong. Please only enter 1 or 2");
+                }
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"An Error Occurred: {ex}");
+            }
+            finally
+            {
+                Console.WriteLine("Program Finished - Press Enter");
+                Console.ReadLine();
+            }
+
+
+        }            
     }
 }
